# APISpecification.yaml
openapi: '3.0.0'
info:
  title: Campus hiring tool
  description: Using this tool you can run a campus hiring drive
  version: "1.0.0"
  contact:
    name: ClearThoughts
    url: http://clearthoughtstechnologies.in
    email: clearthoughtstechnologies@gmail.com
servers:
  - url: http://localhost:8080/v1
paths:
...
  /company:
    post:
      description: create new company entry
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Company'
            example:
              name: ClearThought
              email: clearthoughtstechnologies@gmail.com
              phone: 9980053231
              username: clearthoughts
              password: clear123
              addr1: 2nd Cross
              addr2: Kalyan Nagar
              city: Bangalore
              state: Karnataka
              pin: 560032
      responses:
        '201':
          description: A newly created company
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
              example: 
                _id: 123456
                name: ClearThought
                email: clearthoughtstechnologies@gmail.com
                phone: 9980053231
                username: clearthoughts
                addr1: 2nd Cross
                addr2: Kalyan Nagar
                city: Bangalore
                state: Karnataka
                pin: 560032
    
    put:
      description: modify company
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Company'
            example: 
              _id: 123456
              name: ClearThought
              email: clearthoughtstechnologies@gmail.com
              phone: 9980053232
              username: clearthoughts
              addr1: 2nd Cross
              addr2: Kalyan Nagar
              city: Bangalore
              state: Karnataka
              pin: 560032
      responses:
        '200':
          description: Modified company details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
              example: 
                _id: 123456
                name: ClearThought
                email: clearthoughtstechnologies@gmail.com
                phone: 9980053232
                username: clearthoughts
                addr1: 2nd Cross
                addr2: Kalyan Nagar
                city: Bangalore
                state: Karnataka
                pin: 560032
          
    get:
      description: Get all company listing
      parameters: 
        - name: pageNum
          in: query
          description: page number
          schema: 
            type: integer
        
        - name: noOfElements
          in: query
          description: No of elements in page
          schema:
            type: integer
            
      responses:
        '200':
          description: List of company object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
              example: 
                company1:
                  _id: 123456
                  name: ClearThought
                  email: clearthoughtstechnologies@gmail.com
                  phone: 9980053231
                  username: clearthoughts
                  addr1: 2nd Cross
                  addr2: Kalyan Nagar
                  city: Bangalore
                  state: Karnataka
                  pin: 560032
                
                company2: 
                  _id: 345678
                  name: ClearThought2
                  email: clearthoughtstechnologies2@gmail.com
                  phone: 8880053242
                  username: clearthoughts
                  addr1: 17 Cross
                  addr2: JP Nagar
                  city: Bangalore
                  state: Karnataka
                  pin: 560078
  /company/companyId:
    get:
      description: Return specified company object
      responses:
        '200':
          description: specified company details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
              example: 
                _id: 123456
                name: ClearThought
                email: clearthoughtstechnologies@gmail.com
                phone: 9980053232
                username: clearthoughts
                addr1: 2nd Cross
                addr2: Kalyan Nagar
                city: Bangalore
                state: Karnataka
                pin: 560032
        '404':
          $ref: '#/components/schemas/ErrorBadRequest'
        
#------------------------End--------------------------

#-------------------------Common----------------------
...
  /candidate/registrationUUID:
    get:
      description: This url will be used to fetch candidate basic information. While candidate open the mailed link and trying to register himself. Frontend will fetch stored Candidate info by this URL. registrationUUID will be a variable and dynamic one and that is sent by mail.
      responses:
        '200':
          description: The unregistered TPO detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidateUnregistered'
              example: 
                _id: can123rd
                email: abhi@gmail.com
                phone: 9980053231
                registrationuuid: aedre123wers
                driveid: driveid123
                registered: false
                organization: IIIT Bangalore
                tpoid: tpo123rd

        '404':
          $ref: '#/components/schemas/ErrorBadRequest'
  /candidate:
    post:
      description: Register a new Candidate
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Candidate'
            example: 
              name: Abhijeet Bannerjee
              email: abhi@gmail.com
              phone: 9980053231
              username: canuser
              password: abc123
              addr1: 2nd Cross
              addr2: Kalyan Nagar
              City: Bangalore
              state: Karnataka
              pin: 560032
              registrationuuid: aedre123wers
              driveid: driveid123
              organization: IIIT Bangalore
              tpoid: tpo123rd
              runningyear: 4
              markspercentage: 84
              cvid: abhijeet_cv_12345
      responses:
        '201':
          description: The newly registered candidate detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Candidate'
              example: 
                _id: can123rd
                name: Abhijeet Bannerjee
                email: abhi@gmail.com
                phone: 9980053231
                username: canuser
                addr1: 2nd Cross
                addr2: Kalyan Nagar
                City: Bangalore
                state: Karnataka
                pin: 560032
                registrationuuid: aedre123wers
                driveid: driveid123
                organization: IIIT Bangalore
                tpoid: tpo123rd
                runningyear: 4
                markspercentage: 84
                cvid: abhijeet_cv_12345
    
    put:
      description: Modify a candidate
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Candidate'
            example: 
                _id: can123rd
                name: Abhijeet Bannerjee
                email: abhi@gmail.com
                phone: 9980053231
                username: canuser
                addr1: 2nd Cross
                addr2: Kalyan Nagar
                City: Bangalore
                state: Karnataka
                pin: 560032
                registrationuuid: aedre123wers
                driveid: driveid123
                organization: IIIT Bangalore
                tpoid: tpo123rd
                runningyear: 4
                markspercentage: 84
                cvid: abhijeet_cv_12345
      responses:
        '200':
          description: The candidate detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Candidate'
              example: 
                _id: can123rd
                name: Abhijeet Bannerjee
                email: abhi@gmail.com
                phone: 9980053231
                username: canuser
                addr1: 2nd Cross
                addr2: Kalyan Nagar
                City: Bangalore
                state: Karnataka
                pin: 560032
                registrationuuid: aedre123wers
                driveid: driveid123
                organization: IIIT Bangalore
                tpoid: tpo123rd
                runningyear: 4
                markspercentage: 84
                cvid: abhijeet_cv_12345
          
    get:
      description: Get all Candidate listing
      parameters:
        - name: pageNum
          in: query
          description: page number
          schema: 
            type: integer
        
        - name: noOfElements
          in: query
          description: No of elements in page
          schema:
            type: integer
            
      responses:
        '200':
          description: List of candidate object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Candidate'
              example: 
                candidate1:
                  _id: can123rd
                  name: Abhijeet Bannerjee
                  email: abhi@gmail.com
                  phone: 9980053231
                  username: canuser
                  addr1: 2nd Cross
                  addr2: Kalyan Nagar
                  City: Bangalore
                  state: Karnataka
                  pin: 560032
                  registrationuuid: aedre123wers
                  driveid: driveid123
                  organization: IIIT Bangalore
                  tpoid: tpo123rd
                  runningyear: 4
                  markspercentage: 84
                  cvid: abhijeet_cv_12345
                
                candidate2: 
                  _id: can456xc
                  name: Dipak Pratap
                  email: dipak@gmail.com
                  phone: 9980053231
                  username: canuser
                  addr1: 2nd Cross
                  addr2: Kalyan Nagar
                  City: Bangalore
                  state: Karnataka
                  pin: 560032
                  registrationuuid: aedre123wers
                  driveid: driveid123
                  organization: IIIT Bangalore
                  tpoid: tpo123rd
                  runningyear: 4
                  markspercentage: 84
                  cvid: abhijeet_cv_12345
  /candidate/CandidateId:
    get:
      description: Return specified Candidate object
      responses:
        '200':
          description: The candidate detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Candidate'
              example: 
                _id: can123rd
                name: Abhijeet Bannerjee
                email: abhi@gmail.com
                phone: 9980053231
                username: canuser
                addr1: 2nd Cross
                addr2: Kalyan Nagar
                City: Bangalore
                state: Karnataka
                pin: 560032
                registrationuuid: aedre123wers
                driveid: driveid123
                organization: IIIT Bangalore
                tpoid: tpo123rd
                runningyear: 4
                markspercentage: 84
                cvid: abhijeet_cv_12345
        '404':
          $ref: '#/components/schemas/ErrorBadRequest'
#-------------------------------------------------------------------------
components:
  schemas:
    Company:
      type: object
      properties:
        _id:
          type: string
          example: abcd1234hsc
        name:
          type: string
          example: ClearThought
        email:
          type: string
          example: clearthoughtstechnologies@gmail.com
        phone:
          type: number
          example: 9980053231
        username:
          type: string
          example: clearthoughts
        password:
          type: string
          example: clear123
        addr1:
          type: string
          example: 2nd Cross
        addr2:
          type: string
          example: Kalyan Nagar
        city: 
          type: string
          example: Bangalore
        state:
          type: string
          example: Karnataka
        pin:
          type: number
          example: 560032
...
    Candidate:
      type: object
      properties:
        _id:
          type: string
          example: can123rd
        name:
          type: string
          example: Abhijeet Bannerjee
        email:
          type: string
          example: abhi@gmail.com
        phone:
          type: number
          example: 9980053231
        username:
          type: string
          example: canuser
        password:
          type: string
          example: abc123
        addr1:
          type: string
          example: 2nd Cross
        addr2:
          type: string
          example: Kalyan Nagar
        city: 
          type: string
          example: Bangalore
        state:
          type: string
          example: Karnataka
        pin:
          type: integer
          example: 560032
        registrationuuid:
          type: string
          example: aedre123wers
        driveid:
          type: string
          example: driveid123
        registered:
          type: string
          example: false
        organization:
          type: string
          example: IIIT Bangalore
        tpoid:
          type: string
          example: tpo123rd
        runningyear:
          type: integer
          example: 4
        markspercentage:
          type: integer
          example: 84
        cvid:
          type: string
          example: abhijeet_cv_12345
    CandidateUnregistered:
      type: object
      properties:
        _id:
          type: string
          example: can123rd
        email:
          type: string
          example: abhi@gmail.com
        phone:
          type: number
          example: 9980053231
        registrationuuid:
          type: string
          example: aedre123wers
        driveid:
          type: string
          example: driveid123
        registered:
          type: string
          example: false
        organization:
          type: string
          example: IIIT Bangalore
        tpoid:
          type: string
          example: tpo123rd
    ErrorBadRequest:
      type: object
      properties:
        code:
          type: integer
          example: 400
        message:
          type: string
          example: Bad request
...